+++ dynsys/__init__.py
@@ -1,5 +1,5 @@
 #
 
-__version__ = "1.4.1"
+__version__ = "1.4.1_dev"
 
 __author__ = "Richard Hollamby (RIHY), COWI UK Bridge"